mapeamento de nomes de tabela worldCup

+++ src/DAO/DAO.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface DAO<T> {
-    Optional<T> get(long id);
+    Optional<T> get(int id);
 
     List<T> getAll();
